+++ src/main/java/roomescape/repository/ReservationCollectionRepository.java
@@ -1,10 +1,9 @@
-package roomescape.persistence;
+package roomescape.repository;
 
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
 
 import java.time.LocalDate;
-import java.time.LocalTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
