+++ src/main/java/roomescape/repository/ReservationRepository.java
@@ -1,4 +1,4 @@
-package roomescape.persistence;
+package roomescape.repository;
 
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
