+++ src/test/java/roomescape/controller/AdminControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.presentation;
+package roomescape.controller;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
