+++ src/test/java/roomescape/controller/ControllerTest.java
@@ -1,12 +1,12 @@
-package roomescape.presentation;
+package roomescape.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.web.servlet.MockMvc;
-import roomescape.business.ReservationService;
-import roomescape.business.ReservationTimeService;
+import roomescape.service.ReservationService;
+import roomescape.service.ReservationTimeService;
 
 @WebMvcTest
 public abstract class ControllerTest {
