+++ src/test/java/roomescape/repository/ReservationRepositoryTest.java
@@ -1,4 +1,4 @@
-package roomescape.persistence;
+package roomescape.repository;
 
 interface ReservationRepositoryTest {
